Correctly parse IRC messages
A parsed IRC message was returned as a three-element Array. The trail
was part of the last element (the parameters Array). It was impossible
to tell if the last element from the parameters Array was the trail or
simply the last parameter.

Parsed IRC messages are now returned as a four-element Array. The trail
has been moved out of the parameters Array.